Scout generator module
The module exports an object with a single method:
`generate(options[, callback]) -> Promise(scoutSource)`.

Provide an array of modules as `options.appModules` to specify what to bundle
into the scout. The scout will always include `lib/kernel`, which sets up
essential scout features like the `BV` global symbol.

Unfortunately, webpack doesn't seem to support the specific usecase we have for
a set of modules that should be bundled and dynamically required. We can't use
contexts because we don't know where in the filesystem to look for the
application modules.

Instead, we use a loader to inject a static require, wrapped in a `try`/`catch`
block, for each application module.

[fix #1]
[fix #7]