application module
Allows an application to register itself, which creates a new
property on `window.BV` using the name of the application. The
value of the property is an instance of an `Application`, which
provides:

- a `render` method, which queues its calls until the "real"
application calls the `processQueue` method.
- a `processQueue` method, which allows the "real" application to
provide a handler for processing any queued calls to `render`. The
handler is also used to override the definition of the `render`
method for future calls. This method also deletes the render queue.

This addresses #16, though not exactly as expected.

To run the tests:

```sh
node_modules/.bin/webpack-dev-server --content-base test/fixtures/browser/ --config test/fixtures/browser/unit-tests.webpack.config.js
open http://localhost:8080
```

[close #18]