Remove old IE ordering hack [fix #7]
Re-reading the blog post[1] describing the motivation for the hack, the problem
it addresses, that a second script may execute between the loaded script and the
callback, is not one we are worried about.

While we do expect the callback to always execute after the loaded script, we
do not guarantee that the callback executes before other code.

[1]: http://web.archive.org/web/20130525111405/http://jaubourg.net/2010/07/loading-script-as-onclick-handler-of.html

[close #16]